+++ src/imio/smartweb/core/contents/pages/pages.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from plone.dexterity.content import Container
-from Products.CMFPlone.interfaces import IHideFromBreadcrumbs
 from zope.interface import Interface
 from zope.interface import implementer
 
@@ -10,11 +9,9 @@ class IPages(Interface):
     """Shared base marker interface and schema for Pages"""
 
 
-class IDefaultPages(IHideFromBreadcrumbs):
+class IDefaultPages(Interface):
     """
     Marker interfaces for default Pages in Element view
-    Inheritance from IHideFromBreadcrumbs allows to automatically remove (hide)
-    default pages from breadcrumbs.
     """